refactor: migrate useTxStatusMessage to shared hooks and remove unused component
