+++ src/app/user/page.tsx
@@ -6,7 +6,6 @@ import { BalancesSection } from '@/app/user/Balances/BalancesSection'
 import { CommunitiesSection } from '@/app/user/Communities/CommunitiesSection'
 import { DelegationSection } from '@/app/user/Delegation'
 import { UnderlineTabs, BaseTab } from '@/components/Tabs'
-import { useTxStatusMessage } from '@/components/TxStatusMessage'
 import { usePathname, useSearchParams, useRouter } from 'next/navigation'
 import { useAccount } from 'wagmi'
 import { HeroSection } from './HeroSection'
