--- src/components/TxStatusMessage/index.ts
@@ -1,1 +1,0 @@
-export * from './TxStatusMessage'
